requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.15,<3.17
Pillow>=11.0,<12.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
opencv-python-headless>=4.10,<5.0
